coding #5 registering the category services

+++ OnlineStore/remote/src/main/groovy/mx/ipn/dsd/model/CategoryInteface.groovy
@@ -1,7 +1,7 @@
 package mx.ipn.dsd.model
 
 interface CategoryInterface{
-  Integer getId()
+  Long getId()
   String getName()
   String getDescription()
 }
